start impl label

+++ src/main/java/com/crudjdbc/app/repository/LabelRepository.java
@@ -3,5 +3,5 @@ package com.crudjdbc.app.repository;
 
 import com.crudjdbc.app.model.Label;
 
-public interface LabelRepository extends GenericRepository<Label, Long>{
+public interface LabelRepository extends GenericRepository<Label, Integer>{
 }
