+++ src/main/java/com/crudjdbc/app/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.crudjdbc.app.repository;
 
 import com.crudjdbc.app.model.Post;
 
-public interface PostRepository extends GenericRepository<Post, Long>{
+public interface PostRepository extends GenericRepository<Post, Integer>{
 }
