+++ src/main/java/com/crudjdbc/app/repository/WriterRepository.java
@@ -3,5 +3,5 @@ package com.crudjdbc.app.repository;
 
 import com.crudjdbc.app.model.Writer;
 
-public interface WriterRepository extends GenericRepository<Writer, Long> {
+public interface WriterRepository extends GenericRepository<Writer, Integer> {
 }
